OpenColibri/requirements.txt
Django==1.4.3
psycopg2==2.4.6
django-allauth==0.9.0
dj-database-url==0.2.1
wsgiref==0.1.2
boto==2.6.0
django-storages==1.1.6
#django-athumb==2.0
django-imagekit==2.0.2
django-crispy-forms
django-tastypie==0.9.12
django-ajax-validation==0.1.3
django-ratings
django-countries
django-fluent-comments
PIL==1.1.7
django-voting
git+http://github.com/duointeractive/pial.git
-e git+https://github.com/toastdriven/django-haystack.git#egg=django_haystack-dev
#pyelasticsearch==0.2
-e git+https://github.com/rhec/pyelasticsearch.git#egg=pyelasticsearch
django-tagging
#pillow
South
Markdown<2.3.0
django-sekizai<0.7
django-mptt>=0.5.3
sorl-thumbnail
wiki
#django-celery-with-redis
xlrd
jsonpickle
django-rosetta
gunicorn==0.17.2
simplejson
unicodecsv
jsonfield
#django_evolution
django-articles

